implemented collission detection.

+++ src/server/Player.java
@@ -2,7 +2,7 @@ package server;
 
 public class Player extends Entity{
 
-	public Player(int id, int x, int y) {
+	public Player(int id, float x, float y) {
 		super(id, x, y);
 	}
